# docs/api/api_v3/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Koudaisai Portal API v3",
    "version": "0.1.0"
  },
  "paths": {
...
    "/documents": {
      "get": {
        "tags": [
          "documents"
        ],
        "description": "Get all documents visible to the caller.",
        "operationId": "get_documents",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentRead"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "documents"
        ],
        "description": "Create a document.",
        "operationId": "post_document",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRead"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Invalid document"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/documents/by-category": {
      "get": {
        "tags": [
          "documents"
        ],
        "description": "Get documents visible to the caller, grouped by category.",
        "operationId": "get_documents_by_category",
        "parameters": [
          {
            "name": "include_empty_categories",
            "in": "query",
            "description": "真のとき、ドキュメントが無いカテゴリも空リストで含める。",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentsByCategoryEntry"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/documents/{id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "description": "Get a document by id.",
        "operationId": "get_document",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRead"
                }
              }
            }
          },
          "404": {
            "description": "Document not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "documents"
        ],
        "description": "Delete a document by id.",
        "operationId": "delete_document",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Document deleted"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Document not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "patch": {
        "tags": [
          "documents"
        ],
        "description": "Edit a document by id.",
        "operationId": "patch_document",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Document updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRead"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Document not found"
          },
          "422": {
            "description": "Invalid document"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DocumentCategoryRead": {
        "type": "object",
        "required": [
          "id",
          "created_at",
          "updated_at",
          "title"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "emoji": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "DocumentCreate": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DocumentFormat"
          },
          {
            "type": "object",
            "required": [
              "title",
              "targets"
            ],
            "properties": {
              "category": {
                "type": [
                  "string",
                  "null"
                ],
                "format": "uuid"
              },
              "targets": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "title": {
                "type": "string"
              }
            }
          }
        ]
      },
      "DocumentFormat": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "content",
              "format"
            ],
            "properties": {
              "content": {
                "type": "string"
              },
              "format": {
                "type": "string",
                "enum": [
                  "markdown"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "file_key",
              "file_name",
              "format"
            ],
            "properties": {
              "file_key": {
                "type": "string"
              },
              "file_name": {
                "type": "string"
              },
              "format": {
                "type": "string",
                "enum": [
                  "pdf"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "file_key",
              "file_name",
              "format"
            ],
            "properties": {
              "file_key": {
                "type": "string"
              },
              "file_name": {
                "type": "string"
              },
              "format": {
                "type": "string",
                "enum": [
                  "misc"
                ]
              }
            }
          }
        ]
      },
      "DocumentRead": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DocumentFormat"
          },
          {
            "type": "object",
            "required": [
              "id",
              "created_at",
              "updated_at",
              "created_by",
              "title",
              "targets"
            ],
            "properties": {
              "category": {
                "type": [
                  "string",
                  "null"
                ],
                "format": "uuid"
              },
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "created_by": {
                "type": "string",
                "format": "uuid"
              },
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "targets": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "title": {
                "type": "string"
              },
              "updated_at": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        ]
      },
      "DocumentUpdate": {
        "type": "object",
        "properties": {
          "category": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "format": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DocumentFormat"
              }
            ]
          },
          "targets": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "DocumentsByCategoryEntry": {
        "type": "object",
        "description": "`GET /documents/by-category` の 1 カテゴリ分のエントリ。\n`category` が `None` のときは未分類ドキュメントを表す。",
        "required": [
          "documents"
        ],
        "properties": {
          "category": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DocumentCategoryRead"
              }
            ]
          },
          "documents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentRead"
            }
          }
        }
      },
...
    }
  }
}